Move microsyntaxes parser to alfa-parser
