+++ packages/alfa-dom/test/tables/parse-attribute.spec.ts
@@ -1,12 +1,11 @@
 import { Map } from "@siteimprove/alfa-map";
 import { None, Some } from "@siteimprove/alfa-option";
-import { Parser } from "@siteimprove/alfa-parser";
+import { EnumeratedValueError, Parser } from "@siteimprove/alfa-parser";
 import { Err, Ok } from "@siteimprove/alfa-result";
 import { test } from "@siteimprove/alfa-test";
 
 import { Attribute, Element } from "../../src";
 import {
-  EnumeratedValueError,
   parseAttribute,
   parseEnumeratedAttribute,
 } from "../../src/tables/helpers";
