+++ packages/alfa-parser/src/index.ts
@@ -1,1 +1,2 @@
+export * from "./microsyntaxes";
 export * from "./parser";
